[ISSUE-806][REFACTOR] Remove ResourceConsumption out of ControlMessage
