+++ common/src/main/scala/org/apache/celeborn/common/quota/Quota.scala
@@ -19,7 +19,6 @@ package org.apache.celeborn.common.quota
 
 import org.apache.celeborn.common.identity.UserIdentifier
 import org.apache.celeborn.common.internal.Logging
-import org.apache.celeborn.common.protocol.message.ControlMessages.ResourceConsumption
 import org.apache.celeborn.common.util.Utils
 
 case class Quota(
